pyproject.toml
[project]
name = "secure-locker-api"
version = "0.1.0"
description = "A secure data locker API built with FastAPI"
readme = "README.md"
requires-python = ">=3.13.3"

dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "sqlalchemy",
    "pydantic",
    "python-dotenv",
    "passlib[bcrypt]",
    "cryptography",
    "python-jose[cryptography]",
]
